backend/requirements.txt
aistudio-sdk==0.3.6
annotated-types==0.7.0
anyio==4.10.0
azure-ai-documentintelligence==1.0.2
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.35.1
bce-python-sdk==0.9.45
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
lazy_loader==0.4
MarkupSafe==3.0.2
msrest==0.7.1
num2words==0.5.14
numpy==2.2.6
oauthlib==3.3.1
openai==1.100.2
opt-einsum==3.3.0
orjson==3.11.2
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
pipdeptree==2.28.0
pluggy==1.6.0
prettytable==3.16.0
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pypdfium2==4.30.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
zstandard==0.24.0
